truncate decimals util
