+++ packages/renderer/src/utils/TextUtils.ts
@@ -3,7 +3,7 @@
 
 import { chainCurrency, chainUnits } from '@ren/config/chains';
 import { formatDistanceToNow } from 'date-fns';
-import { planckToUnit } from '@w3ux/utils';
+import { planckToUnit, rmCommas } from '@w3ux/utils';
 import type { ChainID } from '@polkadot-live/types/chains';
 import type {
   NominationPoolCommission,
@@ -98,6 +98,35 @@ export const getNominationPoolStateText = (
     : `Current state is ${curState}.`;
 
 /**
+ * @name truncateDecimals
+ * @summary Truncate decimal string to decimal places.
+ */
+export const truncateDecimals = (value: string, decimals: number): string => {
+  const [intPart, decimalPart] = value.split('.');
+  return !decimalPart
+    ? intPart
+    : `${intPart}.${decimalPart.slice(0, decimals)}`;
+};
+
+/**
+ * @name formatChainUnits
+ * @summary Get readable chain units for rendering.
+ */
+export const formatChainUnits = (units: string, chainId: ChainID) => {
+  // Include regex to remove trailing zeros after decimal point.
+  const asUnit: string = planckToUnit(
+    BigInt(rmCommas(units)),
+    chainUnits(chainId)
+  );
+
+  const formatted: string = truncateDecimals(asUnit, 2)
+    .replace(/(\.\d*?[1-9])0+|\.0*$/, '$1')
+    .replace(/\B(?=(\d{3})+(?!\d))/g, ',');
+
+  return `${formatted} ${chainCurrency(chainId)}`;
+};
+
+/**
  * @name getBalanceText
  * @summary Text to render for transfer events.
  */
@@ -105,7 +134,9 @@ export const getBalanceText = (balance: bigint, chainId: ChainID): string => {
   const asUnit = planckToUnit(balance, chainUnits(chainId));
   const regexA = /\.0+$/; // Remove trailing zeros after a decimal point.
   const regexB = /\B(?=(\d{3})+(?!\d))/g; // Insert commas as thousand separators.
-  const formatted: string = asUnit.replace(regexA, '').replace(regexB, ',');
+  const formatted: string = truncateDecimals(asUnit, 3)
+    .replace(regexA, '')
+    .replace(regexB, ',');
   return `${formatted} ${chainCurrency(chainId)}`;
 };
